+++ MSTest/test_board_utility.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <vector>
 
-std::vector<std::vector<double>> rotateClockwise(
-	std::vector<std::vector<double>> board);
+std::vector<std::vector<int>> rotateClockwise(
+	std::vector<std::vector<int>> board);
